fyers_data_pipeline/requirements.txt
fyers-apiv3
pandas
pyarrow
fastparquet
sqlalchemy
psycopg2-binary
python-dotenv
pyyaml
pytest
tqdm
pyotp
requests
